[Refactor/DetailPage] 삭제 컴포넌트 분리
