+++ src/components/product-detail-page/pdp-right/PdpRight.tsx
@@ -6,8 +6,8 @@ import Tags from "./elements/Tag/Tags";
 import Title from "./elements/SmallUI/Title";
 import UserInfo from "./elements/UserInfo";
 import Location from "./elements/Location";
-import ChatButton from "./elements/ChatButton";
-import ParticipateButton from "./elements/ParticipateButton";
+import ChatButton from "./elements/SmallUI/ChatButton";
+import ParticipateButton from "./elements/SmallUI/ParticipateButton";
 import ExchangeItems from "./elements/EachDetail/ExchangeItems";
 import { useSelector } from "react-redux";
 import { RootState } from "@/store/store";
