+++ src/components/product-detail-page/pdp-right/elements/SmallUI/ChatButton.tsx
@@ -1,7 +1,7 @@
-'use client';
+"use client";
 
-import Link from 'next/link';
-import styled from 'styled-components';
+import Link from "next/link";
+import styled from "styled-components";
 
 const LinkCss = styled(Link)`
   background-color: var(--background-color);
@@ -19,5 +19,5 @@ const LinkCss = styled(Link)`
 
 // 채팅하기 버튼
 export default function ChatButton() {
-  return <LinkCss href='#'>채팅하기</LinkCss>;
+  return <LinkCss href="#">채팅하기</LinkCss>;
 }
