+++ src/components/product-detail-page/pdp-right/elements/SmallUI/ParticipateButton.tsx
